New POIPage created
Point Of Interest Page will display a mapp of all the ports and fishing spots in a given radius from the searched city.
All the components for this page is contained in this directory.
POIPage.js : Main Page component
POIContainer.js : Contains a given type of Point of Interest
POI.js : Component for each individual point of Interest
POIPage.css : Styling for the whole page.